Move WebSocket implementation to io.vertx.core.http.impl.websocket package
