+++ vertx-core/src/main/java/io/vertx/core/http/impl/WebSocketGroup.java
@@ -15,6 +15,7 @@ import io.vertx.core.http.WebSocket;
 import io.vertx.core.http.WebSocketClientOptions;
 import io.vertx.core.http.WebSocketConnectOptions;
 import io.vertx.core.http.impl.http1x.Http1xClientConnection;
+import io.vertx.core.http.impl.websocket.WebSocketImpl;
 import io.vertx.core.internal.ContextInternal;
 import io.vertx.core.internal.PromiseInternal;
 import io.vertx.core.internal.resource.ManagedResource;
