+++ vertx-core/src/main/java/io/vertx/core/http/impl/http1x/Http1xClientConnection.java
@@ -40,6 +40,9 @@ import io.vertx.core.http.impl.*;
 import io.vertx.core.http.impl.HttpRequestHead;
 import io.vertx.core.http.impl.headers.HeadersAdaptor;
 import io.vertx.core.http.impl.headers.Http1xHeaders;
+import io.vertx.core.http.impl.websocket.WebSocketConnectionImpl;
+import io.vertx.core.http.impl.websocket.WebSocketHandshakeInboundHandler;
+import io.vertx.core.http.impl.websocket.WebSocketImpl;
 import io.vertx.core.internal.ContextInternal;
 import io.vertx.core.internal.PromiseInternal;
 import io.vertx.core.internal.buffer.BufferInternal;
