+++ vertx-core/src/main/java/io/vertx/core/http/impl/http1x/Http1xServerConnection.java
@@ -30,6 +30,7 @@ import io.vertx.core.*;
 import io.vertx.core.buffer.Buffer;
 import io.vertx.core.http.ServerWebSocketHandshake;
 import io.vertx.core.http.impl.*;
+import io.vertx.core.http.impl.websocket.ServerWebSocketHandshaker;
 import io.vertx.core.internal.buffer.BufferInternal;
 import io.vertx.core.http.HttpServerOptions;
 import io.vertx.core.http.HttpServerRequest;
