+++ vertx-core/src/main/java/io/vertx/core/http/impl/websocket/ServerWebSocketImpl.java
@@ -9,7 +9,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  */
 
-package io.vertx.core.http.impl;
+package io.vertx.core.http.impl.websocket;
 
 import io.vertx.core.http.ServerWebSocket;
 import io.vertx.core.http.impl.http1x.Http1xServerRequest;
