+++ vertx-core/src/main/java/io/vertx/core/http/impl/websocket/WebSocketFrameInternal.java
@@ -9,7 +9,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  */
 
-package io.vertx.core.http.impl.ws;
+package io.vertx.core.http.impl.websocket;
 
 import io.netty.buffer.ByteBuf;
 import io.vertx.core.http.WebSocketFrame;
