+++ vertx-core/src/main/java/io/vertx/core/http/impl/websocket/WebSocketImpl.java
@@ -9,12 +9,11 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  */
 
-package io.vertx.core.http.impl;
+package io.vertx.core.http.impl.websocket;
 
 import io.vertx.core.Handler;
 import io.vertx.core.http.WebSocket;
 import io.vertx.core.internal.ContextInternal;
-import io.vertx.core.net.impl.VertxConnection;
 
 /**
  * @author <a href="http://tfox.org">Tim Fox</a>
