+++ vertx-core/src/test/java/io/vertx/tests/http/WebSocketTest.java
@@ -40,7 +40,7 @@ import io.vertx.core.http.WebSocketVersion;
 import io.vertx.core.http.impl.http1x.Http1xClientConnection;
 import io.vertx.core.http.impl.http1x.Http1xServerConnection;
 import io.vertx.core.internal.http.WebSocketInternal;
-import io.vertx.core.http.impl.ws.WebSocketFrameImpl;
+import io.vertx.core.http.impl.websocket.WebSocketFrameImpl;
 import io.vertx.core.internal.VertxInternal;
 import io.vertx.core.internal.buffer.BufferInternal;
 import io.vertx.core.internal.net.NetSocketInternal;
